refactor: all sidebars logic/display  in sidebars.rs
